pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yt2bili"
version = "0.1.0"
description = "YouTube channel subscription auto-repost to Bilibili"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "yt-dlp>=2024.0",
    "feedparser>=6.0",
    "pysubs2>=1.6",
    "srt>=3.5",
    "bilibili-api-python>=16.0",
    "apscheduler>=3.10",
    "nicegui>=1.4",
    "typer>=0.9",
    "rich>=13.0",
    "httpx>=0.25",
    "structlog>=23.0",
]

[project.optional-dependencies]
whisper = ["faster-whisper>=0.9"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
yt2bili = "yt2bili.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/yt2bili"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
